+++ src/actions/removeNote.ts
