+++ src/index.ts
@@ -5,7 +5,7 @@ dotenv.config()
 import { init } from "./init"
 import { dates } from "./dates"
 import { counters } from "./counts"
-import { removeNote, newNote, updateNote } from "./notes"
+import { newNote, removeNote, updateNote } from "./actions"
 
 function main() {
     const program = new Command()
